+++ src/graysvr/CWorld.cpp
@@ -2230,7 +2230,7 @@ void CWorld::Speak( const CObjBaseTemplate * pSrc, LPCTSTR pszText, HUE_TYPE wHu
 				if ( sTextGhost.IsEmpty() )
 				{
 					sTextGhost = pszText;
-					for ( int i = 0; i < sTextGhost.GetLength(); i++ )
+					for ( size_t i = 0; i < sTextGhost.GetLength(); i++ )
 					{
 						if ( sTextGhost[i] != ' ' &&  sTextGhost[i] != '\t' )
 							sTextGhost[i] = Calc_GetRandVal(2) ? 'O' : 'o';
